refactor(sidebar-overlay): remove cart-sidebar component for now, make PR only about sidebar-overlay
